refactor(settings): move open target controls into system
Move file-opening preferences into System settings, remove the top-level General entry, and retarget settings fallbacks to System so the remaining navigation and routes stay consistent.
