+++ src-tauri/src/commands/windows.rs
@@ -301,14 +301,13 @@ impl Default for CommitWindowState {
 
 fn resolve_settings_route(section: Option<&str>) -> &'static str {
     match section {
-        Some("general") => "settings/general",
         Some("appearance") => "settings/appearance",
         Some("agents") => "settings/agents",
         Some("mcp") => "settings/mcp",
         Some("skills") => "settings/skills",
         Some("shortcuts") => "settings/shortcuts",
         Some("system") => "settings/system",
-        _ => "settings/general",
+        _ => "settings/system",
     }
 }
