--- src/app/settings/general/page.tsx
@@ -1,5 +1,0 @@
-import { GeneralSettings } from "@/components/settings/general-settings"
-
-export default function SettingsGeneralPage() {
-  return <GeneralSettings />
-}
