+++ src/components/settings/settings-shell.tsx
@@ -20,7 +20,6 @@ import {
   PlugZap,
   Server,
   Settings,
-  SlidersHorizontal,
   Sparkles,
 } from "lucide-react"
 import { useTranslations } from "next-intl"
@@ -37,7 +36,6 @@ import { Sheet, SheetContent, SheetTitle } from "@/components/ui/sheet"
 interface SettingsNavItem {
   href: string
   labelKey:
-    | "general"
     | "appearance"
     | "agents"
     | "model_providers"
@@ -54,11 +52,6 @@ interface SettingsNavItem {
 }
 
 const SETTINGS_NAV_ITEMS: SettingsNavItem[] = [
-  {
-    href: "/settings/general",
-    labelKey: "general",
-    icon: SlidersHorizontal,
-  },
   {
     href: "/settings/appearance",
     labelKey: "appearance",
