+++ src/i18n/messages/es.json
@@ -37,7 +37,6 @@
     "title": "Configuración",
     "preferences": "Preferencias",
     "nav": {
-      "general": "General",
       "appearance": "Apariencia",
       "agents": "Agentes",
       "mcp": "MCP",
@@ -50,29 +49,6 @@
       "model_providers": "Proveedores de Modelos",
       "experts": "Expertos",
       "quick_messages": "Mensajes rápidos"
-    }
-  },
-  "GeneralSettings": {
-    "loading": "Cargando...",
-    "sectionTitle": "General",
-    "sectionDescription": "Configura las preferencias generales de la aplicación.",
-    "loadFailed": "Error al cargar: {message}",
-    "openTargetTitle": "Apertura de archivos",
-    "defaultOpenTarget": "Destino predeterminado de apertura",
-    "defaultOpenTargetDescription": "Ubicación predeterminada para abrir archivos y carpetas",
-    "openTargetDesktopOnly": "Abrir archivos en un editor nativo solo está disponible en la app de escritorio.",
-    "openTargetSaveFailed": "No se pudo guardar el destino de apertura: {message}",
-    "webFileOpenMethod": "Apertura de archivos web",
-    "webFileOpenMethodDescription": "Método predeterminado para abrir HTML y otros archivos web",
-    "webFileOpenMethodSaveFailed": "No se pudo guardar la configuración de apertura de archivos web: {message}",
-    "openTargets": {
-      "vscode": "VS Code",
-      "file_manager": "Gestor de archivos",
-      "terminal": "Terminal"
-    },
-    "webFileOpenMethods": {
-      "browser": "Abrir en el navegador",
-      "editor": "Usar destino de apertura predeterminado"
     }
   },
   "AppearanceSettings": {
@@ -122,6 +98,23 @@
     "loading": "Cargando...",
     "sectionTitle": "Administración del sistema",
     "sectionDescription": "Administra el proxy de red, las actualizaciones de la app y las preferencias de idioma.",
+    "openTargetTitle": "Apertura de archivos",
+    "defaultOpenTarget": "Destino predeterminado de apertura",
+    "defaultOpenTargetDescription": "Ubicación predeterminada para abrir archivos y carpetas",
+    "openTargetDesktopOnly": "Abrir archivos en un editor nativo solo está disponible en la app de escritorio.",
+    "openTargetSaveFailed": "No se pudo guardar el destino de apertura: {message}",
+    "webFileOpenMethod": "Apertura de archivos web",
+    "webFileOpenMethodDescription": "Método predeterminado para abrir HTML y otros archivos web",
+    "webFileOpenMethodSaveFailed": "No se pudo guardar la configuración de apertura de archivos web: {message}",
+    "openTargets": {
+      "vscode": "VS Code",
+      "file_manager": "Gestor de archivos",
+      "terminal": "Terminal"
+    },
+    "webFileOpenMethods": {
+      "browser": "Abrir en el navegador",
+      "editor": "Usar destino de apertura predeterminado"
+    },
     "proxyTitle": "Proxy de red",
     "proxyDescription": "Cuando está activado, las solicitudes de red posteriores usarán este proxy preferentemente (incluyendo chat ACP, instalación de agentes y operaciones remotas de Git).",
     "loadFailed": "Error al cargar: {message}",
