+++ src/i18n/messages/fr.json
@@ -37,7 +37,6 @@
     "title": "Paramètres",
     "preferences": "Préférences",
     "nav": {
-      "general": "Général",
       "appearance": "Apparence",
       "agents": "Agents IA",
       "mcp": "MCP",
@@ -50,29 +49,6 @@
       "model_providers": "Fournisseurs de Modèles",
       "experts": "Experts",
       "quick_messages": "Messages rapides"
-    }
-  },
-  "GeneralSettings": {
-    "loading": "Chargement...",
-    "sectionTitle": "Général",
-    "sectionDescription": "Configurez les préférences générales de l’application.",
-    "loadFailed": "Échec du chargement : {message}",
-    "openTargetTitle": "Ouverture de fichiers",
-    "defaultOpenTarget": "Cible d’ouverture par défaut",
-    "defaultOpenTargetDescription": "Emplacement par défaut pour ouvrir les fichiers et dossiers",
-    "openTargetDesktopOnly": "L’ouverture dans un éditeur natif est disponible uniquement dans l’app de bureau.",
-    "openTargetSaveFailed": "Échec de l’enregistrement de la cible d’ouverture : {message}",
-    "webFileOpenMethod": "Ouverture des fichiers web",
-    "webFileOpenMethodDescription": "Méthode par défaut pour ouvrir les fichiers HTML et autres fichiers web",
-    "webFileOpenMethodSaveFailed": "Échec de l’enregistrement de l’ouverture des fichiers web : {message}",
-    "openTargets": {
-      "vscode": "VS Code",
-      "file_manager": "Gestionnaire de fichiers",
-      "terminal": "Terminal"
-    },
-    "webFileOpenMethods": {
-      "browser": "Ouvrir dans le navigateur",
-      "editor": "Utiliser la cible d’ouverture par défaut"
     }
   },
   "AppearanceSettings": {
@@ -122,6 +98,23 @@
     "loading": "Chargement...",
     "sectionTitle": "Gestion du système",
     "sectionDescription": "Gérez le proxy réseau, les mises à jour de l’app et les préférences de langue.",
+    "openTargetTitle": "Ouverture de fichiers",
+    "defaultOpenTarget": "Cible d’ouverture par défaut",
+    "defaultOpenTargetDescription": "Emplacement par défaut pour ouvrir les fichiers et dossiers",
+    "openTargetDesktopOnly": "L’ouverture dans un éditeur natif est disponible uniquement dans l’app de bureau.",
+    "openTargetSaveFailed": "Échec de l’enregistrement de la cible d’ouverture : {message}",
+    "webFileOpenMethod": "Ouverture des fichiers web",
+    "webFileOpenMethodDescription": "Méthode par défaut pour ouvrir les fichiers HTML et autres fichiers web",
+    "webFileOpenMethodSaveFailed": "Échec de l’enregistrement de l’ouverture des fichiers web : {message}",
+    "openTargets": {
+      "vscode": "VS Code",
+      "file_manager": "Gestionnaire de fichiers",
+      "terminal": "Terminal"
+    },
+    "webFileOpenMethods": {
+      "browser": "Ouvrir dans le navigateur",
+      "editor": "Utiliser la cible d’ouverture par défaut"
+    },
     "proxyTitle": "Proxy réseau",
     "proxyDescription": "Lorsqu’il est activé, les requêtes réseau suivantes utilisent ce proxy en priorité (y compris le chat ACP, l’installation d’agents et les opérations Git distantes).",
     "loadFailed": "Échec du chargement : {message}",
