+++ src/i18n/messages/pt.json
@@ -37,7 +37,6 @@
     "title": "Configurações",
     "preferences": "Preferências",
     "nav": {
-      "general": "Geral",
       "appearance": "Aparência",
       "agents": "Agentes",
       "mcp": "MCP",
@@ -50,29 +49,6 @@
       "model_providers": "Provedores de Modelos",
       "experts": "Especialistas",
       "quick_messages": "Mensagens rápidas"
-    }
-  },
-  "GeneralSettings": {
-    "loading": "Carregando...",
-    "sectionTitle": "Geral",
-    "sectionDescription": "Configure as preferências gerais do app.",
-    "loadFailed": "Falha ao carregar: {message}",
-    "openTargetTitle": "Abertura de arquivos",
-    "defaultOpenTarget": "Destino padrão de abertura",
-    "defaultOpenTargetDescription": "Local padrão para abrir arquivos e pastas",
-    "openTargetDesktopOnly": "Abrir arquivos em um editor nativo está disponível apenas no app desktop.",
-    "openTargetSaveFailed": "Falha ao salvar o destino de abertura: {message}",
-    "webFileOpenMethod": "Abertura de arquivos web",
-    "webFileOpenMethodDescription": "Método padrão para abrir HTML e outros arquivos web",
-    "webFileOpenMethodSaveFailed": "Falha ao salvar as configurações de abertura de arquivos web: {message}",
-    "openTargets": {
-      "vscode": "VS Code",
-      "file_manager": "Gerenciador de arquivos",
-      "terminal": "Terminal"
-    },
-    "webFileOpenMethods": {
-      "browser": "Abrir no navegador",
-      "editor": "Usar destino padrão de abertura"
     }
   },
   "AppearanceSettings": {
@@ -122,6 +98,23 @@
     "loading": "Carregando...",
     "sectionTitle": "Gerenciamento do sistema",
     "sectionDescription": "Gerencie proxy de rede, atualizações do app e preferências de idioma.",
+    "openTargetTitle": "Abertura de arquivos",
+    "defaultOpenTarget": "Destino padrão de abertura",
+    "defaultOpenTargetDescription": "Local padrão para abrir arquivos e pastas",
+    "openTargetDesktopOnly": "Abrir arquivos em um editor nativo está disponível apenas no app desktop.",
+    "openTargetSaveFailed": "Falha ao salvar o destino de abertura: {message}",
+    "webFileOpenMethod": "Abertura de arquivos web",
+    "webFileOpenMethodDescription": "Método padrão para abrir HTML e outros arquivos web",
+    "webFileOpenMethodSaveFailed": "Falha ao salvar as configurações de abertura de arquivos web: {message}",
+    "openTargets": {
+      "vscode": "VS Code",
+      "file_manager": "Gerenciador de arquivos",
+      "terminal": "Terminal"
+    },
+    "webFileOpenMethods": {
+      "browser": "Abrir no navegador",
+      "editor": "Usar destino padrão de abertura"
+    },
     "proxyTitle": "Proxy de rede",
     "proxyDescription": "Quando ativado, as solicitações de rede seguintes priorizam este proxy (incluindo chat ACP, instalação de agentes e operações remotas do Git).",
     "loadFailed": "Falha ao carregar: {message}",
