+++ src/i18n/messages/zh-CN.json
@@ -37,7 +37,6 @@
     "title": "设置",
     "preferences": "偏好设置",
     "nav": {
-      "general": "常规",
       "appearance": "外观",
       "agents": "智能体",
       "mcp": "MCP",
@@ -50,29 +49,6 @@
       "model_providers": "模型供应商",
       "experts": "专家",
       "quick_messages": "快捷消息"
-    }
-  },
-  "GeneralSettings": {
-    "loading": "加载中...",
-    "sectionTitle": "常规",
-    "sectionDescription": "配置应用的通用偏好设置。",
-    "loadFailed": "加载失败：{message}",
-    "openTargetTitle": "文件打开",
-    "defaultOpenTarget": "默认打开目标",
-    "defaultOpenTargetDescription": "打开文件和文件夹的默认位置",
-    "openTargetDesktopOnly": "在原生编辑器中打开文件仅桌面应用可用。",
-    "openTargetSaveFailed": "打开目标设置保存失败：{message}",
-    "webFileOpenMethod": "网页文件打开方式",
-    "webFileOpenMethodDescription": "html等文件默认打开方式",
-    "webFileOpenMethodSaveFailed": "网页文件打开方式保存失败：{message}",
-    "openTargets": {
-      "vscode": "VS Code",
-      "file_manager": "文件管理器",
-      "terminal": "终端"
-    },
-    "webFileOpenMethods": {
-      "browser": "在浏览器中打开",
-      "editor": "使用默认打开目标"
     }
   },
   "AppearanceSettings": {
@@ -122,6 +98,23 @@
     "loading": "加载中...",
     "sectionTitle": "系统管理",
     "sectionDescription": "管理网络代理、应用升级与语言偏好。",
+    "openTargetTitle": "文件打开",
+    "defaultOpenTarget": "默认打开目标",
+    "defaultOpenTargetDescription": "打开文件和文件夹的默认位置",
+    "openTargetDesktopOnly": "在原生编辑器中打开文件仅桌面应用可用。",
+    "openTargetSaveFailed": "打开目标设置保存失败：{message}",
+    "webFileOpenMethod": "网页文件打开方式",
+    "webFileOpenMethodDescription": "html等文件默认打开方式",
+    "webFileOpenMethodSaveFailed": "网页文件打开方式保存失败：{message}",
+    "openTargets": {
+      "vscode": "VS Code",
+      "file_manager": "文件管理器",
+      "terminal": "终端"
+    },
+    "webFileOpenMethods": {
+      "browser": "在浏览器中打开",
+      "editor": "使用默认打开目标"
+    },
     "proxyTitle": "网络代理",
     "proxyDescription": "开启后，后续网络请求将优先走该代理（包括 ACP 对话、Agent 安装、Git 远程操作等）。",
     "loadFailed": "加载失败：{message}",
