+++ src/lib/api.ts
@@ -1128,7 +1128,6 @@ export async function openCommitWindow(folderId: number): Promise<void> {
 }
 
 export type SettingsSection =
-  | "general"
   | "appearance"
   | "agents"
   | "mcp"
@@ -1158,7 +1157,7 @@ export async function openSettingsWindow(
       agentType: options?.agentType ?? null,
     }
   )
-  window.open(result.path, `settings-${section ?? "general"}`)
+  window.open(result.path, `settings-${section ?? "system"}`)
 }
 
 export async function openProjectBootWindow(source?: string): Promise<void> {
