+++ src/lib/tauri.ts
@@ -881,7 +881,6 @@ export async function openCommitWindow(folderId: number): Promise<void> {
 }
 
 export type SettingsSection =
-  | "general"
   | "appearance"
   | "agents"
   | "mcp"
